+++ src/sockets/SocketClient.cpp
@@ -1,0 +1,35 @@
+#include "extras/sockets/SocketClient.hpp"
+
+#include <arpa/inet.h>
+#include <stdio.h>
+#include <string.h>
+#include <sys/socket.h>
+#include <unistd.h>
+
+#include <iostream>
+#include <sstream>
+
+using namespace std;
+
+namespace extras {
+
+  SocketClient::SocketClient(const std::string &hostname, int port)
+      : _hostname(hostname) {
+    this->_socket = socket(AF_INET, SOCK_STREAM, 0);
+    SocketException::assertLTZ(_socket, "SocketClient creation error",
+                               __INFO__);
+    this->_serv_addr.sin_family = AF_INET;
+    this->_serv_addr.sin_port = htons(port);
+    // Convert IPv4 and IPv6 addresses from text to binary form
+    auto test = inet_pton(AF_INET, _hostname.c_str(), &_serv_addr.sin_addr);
+    SocketException::assertLTZ(test, "IPv6 addresses not supported", __INFO__);
+    _proxy = new Socket(port, this->_socket);
+  }
+
+  void SocketClient::connect() {
+    auto test = ::connect(this->_socket, (struct sockaddr *)&_serv_addr,
+                          sizeof(_serv_addr));
+    SocketException::assertLTZ(test, "Connection Failed", __INFO__);
+  }
+
+}  // namespace extras