+++ test/mock_Socket.cpp
@@ -1,0 +1,26 @@
+
+#include <iostream>
+
+#include "catch.hpp"
+#include "extras/sockets/SocketClient.hpp"
+#include "fakeit.hpp"
+
+//
+// https://github.com/eranpeer/FakeIt/wiki/Quickstart
+//
+
+using namespace extras;
+using namespace fakeit;
+
+/**
+ * @brief Mock FakeIt AlwaysDo demonstration
+ *
+ */
+SCENARIO("Mock SocketClientInterface: connect", "[SocketInterface]") {
+  Mock<SocketClientInterface> mock;
+  When(Method(mock, connect)).Return();
+
+  SocketClientInterface &i = mock.get();
+  i.connect();
+  Verify(Method(mock, connect));
+}