+++ test/test_SocketServer.cpp
@@ -1,0 +1,27 @@
+#include <cstdint>
+#include <extras/paths.hpp>
+#include <extras/strings.hpp>
+
+#include "catch.hpp"
+#include "extras/sockets/SocketServer.hpp"
+
+using namespace std;
+using namespace extras;
+
+// SCENARIO("Verify SocketServerInterface SocketServer()",
+//          "[SocketServerInterface]") {
+//   int port = 8000;
+//   SocketServer socket(port);
+// }
+
+// SCENARIO("Verify SocketServerInterface accept()", "[SocketServerInterface]")
+// {
+//   int port = 8000;
+//   SocketServer socket(port);
+//   socket.accept();
+// }
+
+// SCENARIO("Verify SocketServerInterface socket_server",
+//          "[SocketServerInterface]") {
+//     system("build/socket_server &");
+// }